The 'SCOPE_USER' scope and 'getResourceOwner' method are now fixed.
After some communication with Twinfield, it became apparent that the 'twf.user' scope was not working, because the 'openid' scope was missing. Unfortunately, this was not stated in their documentation. Furthermore, the problem with the 'getResourceOwner' method is fixed. It can be called only when the 'openid' scope is included in the authorization url. The 'twf.user' scope can also be passed to retrieve more information from the 'getResourceOwner' method.

Changes:
- Added the 'openid' scope as constant.
- Added the 'openid' scope as a default scope, as the default scopes should include those scopes that are necessary to request the details of the resource owner.
- Made sure to allow the 'twf.user' scope again, but only if the 'openid' scope is present.
- Modified the 'ResourceOwner' class to match the data that Twinfield returns.