Add `BookingReferenceInterface` for identifying bookings
Deleting transactions, statements, etc. is done using a reference to
that object. However, the existing `MatchReferenceInterface` does not
suffice for this, since it references a specific line within such an
object.

Therefore `BookingReferenceInterface` was added to unqiuely identify a
booking.

`BookingReferenceInterface` is not extended by `MatchReferenceInterface`
on purpose, since we don't want anyone to accidentally pass a
`MatchReferenceInterface` to a delete method, expecting the specific
line to be deleted, but it actually deletes the entire transaction.
Usage of `BookingReferenceInterface` should be explicit.

Ease of use methods `BookingReference::fromMatchReference()` and
`MatchReference::fromBookingReference()` have been added.